requirements.txt
requests
json
openai
python-dotenv